// mogadishu rentals/types.ts
export interface Message {
  id: string;
  sender: string;
  senderAvatar: string;
  lastMessage: string;
  time: string;
  productTitle: string;
  productImage: string;
  unreadCount: number;
  type: 'Buying' | 'Selling';
}